moved drawer into own component
