+++ client/src/components/Jumbotron/SignUpModal/SignUpModal.js
@@ -6,8 +6,6 @@ import Dialog from "@material-ui/core/Dialog";
 import DialogActions from "@material-ui/core/DialogActions";
 import DialogContent from "@material-ui/core/DialogContent";
 import DialogTitle from "@material-ui/core/DialogTitle";
-import axios from 'axios'
-
 
 const useStyles = makeStyles({
   modal: {
