+++ client/src/components/Navbar/Drawer/index.js
@@ -1,0 +1,1 @@
+export { default } from './Drawer.js'
